requirements.txt
APScheduler==3.6.3
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
decorator==5.0.7
idna==2.10
lxml==4.6.3
python-telegram-bot==13.5
pytz==2021.1
requests==2.25.1
robobrowser==0.5.3
six==1.16.0
soupsieve==2.2.1
tornado==6.1
tzlocal==2.1
urllib3==1.26.4
validators==0.18.2
Werkzeug==0.16.0
